pyproject.toml
[tool.poetry]
name = "scope3_methodology"
version = "0.1.0"
description = ""
authors = ["scope3 <hello@scope3.com>"]
license = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
mypy = "^0.971"
black = { version = "^22.6.0", allow-prereleases = true }
scriv = { extras = ["toml"], version = "^0.16.0" }
types-PyYAML = "^6.0.11"
pytest = "^7.1.2"
coverage = "^6.4.4"
pylint = "^2.15.0"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.scriv]
version = "literal: scope3_methodology/__init__.py: __version__"
format = "md"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
allow_redefinition = false
check_untyped_defs = true
follow_imports = "normal"
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
local_partial_types = true
namespace_packages = true
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = false
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true
